KA-66 Rename IItem,IItemFlags,remove unused export
Rename to Item and ItemFlags as interfaces are only for internal use and
Item(Flags) is major here
